(MAX_PROP_TABLE_SIZE, MAX_OBJECTS): Increase
(Street): Fix the code to determine if the player is going to a
street.  It looks like the old code only worked by coincidence, but
suddenly broke.  From what I can tell, the object numbers happened to
coincide with the property numbers.  That's no longer true, probably
from the CvsId addition.