(bridge): Fix GO EAST blocked message
(bridge): If Jacques or Ralph have the photo anywhere in their
possession, it's cool.

(Proprietor): When showing the photograph, just make sure that it's
somewhere in inventory, not directly so.

(helicopter): Examine should work even if the player somehow is in
something else in the chopper.  Also, search should give a custom
message on failure.

(helicopter): Add Take,Push,Pull handlers

(helicopter): Make transparent, so you can interact (somewhat) with
Carla.

(yoke): Reveal the tires after the take (so don't rely on MOVED to
describe it).

(carla): Allow a description to happen for outside the chopper, but no
actions other than examine.  (Handlers on before and life.)